create a PhotoGrid component
